admin authorization for author's post bug: image not shown for all posts

+++ app/Http/Controllers/Author/PostController.php
@@ -105,6 +105,10 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if($post -> user_id != Auth::id()){
+            Toastr::error('Unauthorized Request', 'Error');
+            return redirect()->back();
+        }
         return view('author.post.show', compact('post'));
     }
 
@@ -116,6 +120,10 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post -> user_id != Auth::id()){
+            Toastr::error('Unauthorized Request', 'Error');
+            return redirect()->back();
+        }
         $categories = Category::all();
         $tags = Tag::all();
         return view ('author.post.edit',compact('categories','tags', 'post'));
@@ -130,6 +138,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if($post -> user_id != Auth::id()){
+            Toastr::error('Unauthorized Request', 'Error');
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required',
             'image' => 'image',
@@ -194,6 +207,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post -> user_id != Auth::id()){
+            Toastr::error('Unauthorized Request', 'Error');
+            return redirect()->back();
+        }
+
         if(Storage::disk('public')->exists('post/'.$post->image))
         {
             Storage::disk('public')->delete('post/'.$post->image);
